New matrices of the SVAR50_4B_QPM
This matrices were obtained from GLS estimation process from the new HEFM scheme.
The data that used to estimate the matrices were the same as those used in the QPM predictions.

obs xlsx file is an excel file that compares the observed data of the new SVAR50_4B_QPM with the observed used in QPM model, so the obs xlsx file is a control file.